modularize app and separate call swagger
